Cambios relevantes realizados en el proyecto:
1. Se convirtio el Frame en un InternalFrame para mostrarlo en el panel
de Vista previa

2. Modificaciones en el StdDraw3D para lograr configurar el frame.

3. Se implementó un nuevo metodo de limpiar la pizarra 3D

+++ Proyecto_Final/src/Logica/Rectangulo.java
@@ -82,7 +82,7 @@ public class Rectangulo extends Prismas {
 
 
 	@Override
-	public float perimetro() {
+	public double perimetro() {
 		
 		float lado1 = 0;
 		float lado2 = 0;
